fix(daemon): case-insensitive reserved-name guard + share realSpawn + lock consent-order test
M1: safeFileName's RESERVED check is now case-insensitive — APFS/HFS+
(macOS default) is case-insensitive, so START.COMMAND/Context.MD
previously slipped the guard yet clobbered the same on-disk reserved
file. M2: extracted realSpawn+SpawnFn (with the pipe-deadlock rationale
comment) into daemon/src/spawn.ts, removing the duplicate in handoff.ts.
M3: locks reject-before-consent ordering in the empty-context test by
asserting requestConsent/run were never called.
